[Gold IV] Title: 좋은 친구, Time: 188 ms, Memory: 5108 KB -BaekjoonHub

+++ 백준/Gold/3078. 좋은 친구/좋은 친구.cc
@@ -3,7 +3,7 @@
 using namespace std;
 int main(void) {
     int n, k;
-    long long cnt = 0;
+    long cnt = 0;
     string t;
     vector<int> name, len_name;
     cin >> n >> k;
